+++ lib/api.js
@@ -13,7 +13,7 @@ class Api {
    * @param {string} options.version - Kubernetes API version
    * @param {string} options.namespace - Default namespace
    */
-  constructor(options, genericTypes) {
+  constructor(options) {
     this.url = options.url;
     this.version = options.version;
     this.path = `/${ options.path }/${ this.version }`;
@@ -39,7 +39,7 @@ class Api {
     //
     // Create resources that live at the root (e.g., /api/v1/nodes)
     //
-    for (const type of genericTypes) {
+    for (const type of options.genericTypes) {
       this[type] = new BaseObject({
         path: `${ this.path }/${ type }`,
         api: this
