+++ lib/core.js
@@ -5,8 +5,6 @@ var aliasResources = require('./common').aliasResources;
 
 class Core extends Api {
   constructor(options) {
-    options.path = 'api';
-    options.version = options.version || 'v1';
     const genericTypes = [
       'componentstatuses',
       'configmaps',
@@ -24,8 +22,12 @@ class Core extends Api {
       'serviceaccounts',
       'services'
     ];
-
-    super(options, genericTypes);
+    options = Object.assign({}, options, {
+      path: 'api',
+      version: options.version || 'v1',
+      genericTypes: genericTypes
+    });
+    super(options);
   }
 }
