+++ lib/extensions.js
@@ -4,8 +4,6 @@ var Api = require('./api');
 
 class ApisExtension extends Api {
   constructor(options) {
-    options.path = 'apis/extensions';
-    options.version = options.version || 'v1beta1';
     const genericTypes = [
       'daemonsets',
       'deployments',
@@ -13,7 +11,12 @@ class ApisExtension extends Api {
       'ingresses',
       'jobs'
     ];
-    super(options, genericTypes);
+    options = Object.assign({}, options, {
+      path: 'apis/extensions',
+      version: options.version || 'v1beta1',
+      genericTypes: genericTypes
+    });
+    super(options);
   }
 }
