+++ lib/index.js
@@ -1,5 +1,5 @@
 module.exports.Core = require('./core');
-module.exports.ApisExtension = require('./apisExtension');
+module.exports.ApisExtension = require('./extensions');
 module.exports.testUtils = {
   aliasResources: require('./common').aliasResources
 };
